requirements-base.txt
--index-url https://pypi.devinfra.sentry.io/simple

beautifulsoup4>=4.7.1
boto3>=1.34.128
botocore>=1.34.8
brotli>=1.1.0
cachetools>=5
celery>=5
click>=8.1
confluent-kafka>=2.3.0
cronsim>=2.6
cssselect>=1.0.3
datadog>=0.49
django-crispy-forms>=1.14.0
django-csp>=3.8
django-pg-zero-downtime-migrations>=0.16
Django>=5.1.7
djangorestframework>=3.16
drf-spectacular>=0.27.2
sentry-forked-email-reply-parser>=0.5.12.post1
google-api-core>=2.19.1
google-auth>=2.29.0
google-cloud-bigtable>=2.26.0
google-cloud-build>=3.24.2
google-cloud-core>=2.4.1
google-cloud-functions>=1.17.0
google-cloud-kms>=2.24.2
google-cloud-pubsub>=2.23.0
google-cloud-spanner>=3.49.0
google-cloud-storage>=2.18.0
googleapis-common-protos>=1.63.2
google-crc32c>=1.6.0
grpc-google-iam-v1>=0.13.1
isodate>=0.6.1
joserfc>=1.0.4
jsonschema>=3.2.0
lxml>=5.3.0
maxminddb>=2.3
mistune>=2.0.3
mmh3>=4.0.0
packaging>=24.1
parsimonious>=0.10.0
petname>=2.6
phonenumberslite>=8.12.32
Pillow>=11.0.0
progressbar2>=3.41.0
protobuf>=5.27.3
proto-plus>=1.25.0
python-rapidjson>=1.4
psutil>=5.9.2
psycopg2-binary>=2.9.10
PyJWT>=2.4.0
pydantic>=1.10.20,<2
python-dateutil>=2.9.0
pymemcache
python-u2flib-server>=5.0.0
fido2>=0.9.2
python3-saml>=1.15.0
PyYAML>=6.0.2
rb>=1.9.0
redis-py-cluster>=2.1.0
redis>=3.4.1
requests-oauthlib>=1.2.0
requests>=2.32.3
# [start] jsonschema format validators
rfc3339-validator>=0.1.2
rfc3986-validator>=0.1.1
# [end] jsonschema format validators
sentry-arroyo>=2.21.0
sentry-kafka-schemas>=1.2.0
sentry-ophio==1.0.0
sentry-protos==0.1.70
sentry-redis-tools>=0.5.0
sentry-relay>=0.9.8
sentry-sdk[http2]>=2.25.1
slack-sdk>=3.27.2
snuba-sdk>=3.0.43
simplejson>=3.17.6
sqlparse>=0.4.4
statsd>=3.3
structlog>=22
symbolic==12.14.1
tiktoken>=0.8.0
tldextract>=5.1.2
toronado>=0.1.0
typing-extensions>=4.9.0
ua-parser>=0.10.0
unidiff>=0.7.4
urllib3[brotli]>=2.2.2
pyuwsgi==2.0.28.post1
zstandard>=0.18.0
sentry-usage-accountant==0.0.10
orjson>=3.10.10

msgpack>=1.1
cryptography>=38.0.3

# Note, grpcio>1.30.0 requires setting GRPC_POLL_STRATEGY=epoll1
# See https://github.com/grpc/grpc/issues/23796 and
# https://github.com/grpc/grpc/blob/v1.35.x/doc/core/grpc-polling-engines.md#polling-engine-implementations-in-grpc
grpcio>=1.67

# not directly used, but provides a speedup for redis
hiredis>=2.3.2

openai==1.3.5

# remove once there are no unmarked transitive dependencies on setuptools
setuptools
